dev * Updated tests.
